improve module organization
