+++ src/Swarm/Game/Scenario/Scoring/CodeSize.hs
@@ -1,3 +1,5 @@
+-- | Types and utilities to compute code size
+-- in terms of textual length and AST.
 module Swarm.Game.Scenario.Scoring.CodeSize where
 
 import Control.Monad (guard)
