+++ src/Swarm/Game/ScenarioInfo.hs
@@ -14,10 +14,6 @@ module Swarm.Game.ScenarioInfo (
   ScenarioInfo (..),
   scenarioPath,
   scenarioStatus,
-  scenarioBestByTime,
-  scenarioBestByTicks,
-  scenarioBestByCharCount,
-  scenarioBestByAstSize,
   CodeSizeDeterminators (CodeSizeDeterminators),
   updateScenarioInfoOnFinish,
   ScenarioInfoPair,
