+++ src/Swarm/TUI/View.hs
@@ -73,9 +73,10 @@ import Swarm.Game.Location
 import Swarm.Game.Recipe
 import Swarm.Game.Robot
 import Swarm.Game.Scenario (scenarioAuthor, scenarioDescription, scenarioName, scenarioObjectives)
+import Swarm.Game.Scenario.Scoring.Best
 import Swarm.Game.Scenario.Scoring.CodeSize
-import Swarm.Game.Scenario.Scoring.Metrics
-import Swarm.Game.Scenario.Scoring.Progress
+import Swarm.Game.Scenario.Scoring.ConcreteMetrics
+import Swarm.Game.Scenario.Scoring.GenericMetrics
 import Swarm.Game.Scenario.Status
 import Swarm.Game.ScenarioInfo (
   ScenarioItem (..),
