+++ swarm.cabal
@@ -109,9 +109,10 @@ library
                       Swarm.Game.Scenario.Objective.Validation
                       Swarm.Game.Scenario.Objective.WinCheck
                       Swarm.Game.Scenario.RobotLookup
+                      Swarm.Game.Scenario.Scoring.Best
                       Swarm.Game.Scenario.Scoring.CodeSize
-                      Swarm.Game.Scenario.Scoring.Metrics
-                      Swarm.Game.Scenario.Scoring.Progress
+                      Swarm.Game.Scenario.Scoring.ConcreteMetrics
+                      Swarm.Game.Scenario.Scoring.GenericMetrics
                       Swarm.Game.Scenario.Status
                       Swarm.Game.Scenario.Style
                       Swarm.Game.Scenario.WorldDescription
