+++ test/unit/TestScoring.hs
@@ -3,20 +3,19 @@
 -- | High score records
 module TestScoring where
 
-import Swarm.Language.Syntax
-import Swarm.Language.Module
-import Swarm.Language.Pipeline
-import System.FilePath ((</>))
-import Data.Time.Calendar.OrdinalDate
 import Data.Text.IO qualified as TIO
+import Data.Time.Calendar.OrdinalDate
 import Data.Time.LocalTime
 import Swarm.Game.Scenario.Scoring.CodeSize
-import Swarm.Game.Scenario.Scoring.Metrics
-import Swarm.Game.Scenario.Scoring.Progress
+import Swarm.Game.Scenario.Scoring.ConcreteMetrics
+import Swarm.Game.Scenario.Scoring.GenericMetrics
 import Swarm.Game.Scenario.Status
+import Swarm.Language.Module
+import Swarm.Language.Pipeline
+import Swarm.Language.Syntax
+import System.FilePath ((</>))
 import Test.Tasty
 import Test.Tasty.HUnit
-
 
 baseTestPath :: FilePath
 baseTestPath = "data/test/language-snippets/code-size"
@@ -25,36 +24,37 @@ testHighScores :: TestTree
 testHighScores =
   testGroup
     "Scoring"
-    [ testGroup "Code size" [
-      compareAstSize 1 "single-move-bare.sw"
-    , compareAstSize 2 "single-move-def.sw"
-    , compareAstSize 3 "single-move-let-with-invocation.sw"
-    , compareAstSize 5 "double-move-let-with-invocation.sw"
-    , compareAstSize 6 "single-move-def-with-invocation.sw"
-    , compareAstSize 8 "double-move-def-with-invocation.sw"
-    , compareAstSize 28 "single-def-two-args-recursive.sw"
-    , compareAstSize 36 "single-def-two-args-recursive-with-invocation.sw"
-    ]
-  , testGroup
-    "Precedence"
     [ testGroup
-        "Single metrics"
-        [ testCase
-            "Attempted long > Attempted short"
-            $ assertEqual "Longer is better for incomplete games" (Metric Attempted (5 :: Int))
-            $ chooseBetter pure (Metric Attempted 3) (Metric Attempted 5)
-        , testCase
-            "Completed short > Completed long"
-            $ assertEqual "Shorter is better for completed games" (Metric Completed (3 :: Int))
-            $ chooseBetter pure (Metric Completed 3) (Metric Completed 5)
+        "Code size"
+        [ compareAstSize 1 "single-move-bare.sw"
+        , compareAstSize 2 "single-move-def.sw"
+        , compareAstSize 3 "single-move-let-with-invocation.sw"
+        , compareAstSize 5 "double-move-let-with-invocation.sw"
+        , compareAstSize 6 "single-move-def-with-invocation.sw"
+        , compareAstSize 8 "double-move-def-with-invocation.sw"
+        , compareAstSize 28 "single-def-two-args-recursive.sw"
+        , compareAstSize 36 "single-def-two-args-recursive-with-invocation.sw"
         ]
     , testGroup
-        "Grouped metrics"
-        [ betterReplTimeAfterCodeSizeRecord
-        , betterCodeWorseTime
+        "Precedence"
+        [ testGroup
+            "Single metrics"
+            [ testCase
+                "Attempted long > Attempted short"
+                $ assertEqual "Longer is better for incomplete games" (Metric Attempted (5 :: Int))
+                $ chooseBetter pure (Metric Attempted 3) (Metric Attempted 5)
+            , testCase
+                "Completed short > Completed long"
+                $ assertEqual "Shorter is better for completed games" (Metric Completed (3 :: Int))
+                $ chooseBetter pure (Metric Completed 3) (Metric Completed 5)
+            ]
+        , testGroup
+            "Grouped metrics"
+            [ betterReplTimeAfterCodeSizeRecord
+            , betterCodeWorseTime
+            ]
         ]
     ]
-  ]
 
 compareAstSize :: Int -> FilePath -> TestTree
 compareAstSize expectedSize path = testCase (unwords ["size of", path]) $ do
@@ -64,7 +64,6 @@ compareAstSize expectedSize path = testCase (unwords ["size of", path]) $ do
     Left y -> assertFailure y
   let actualSize = measureAstSize stx
   assertEqual "incorrect size" expectedSize actualSize
-
 
 betterReplTimeAfterCodeSizeRecord :: TestTree
 betterReplTimeAfterCodeSizeRecord =
